works with Double

+++ HackerRank/Functional_Programming/Easy/Introduction/Evaluating_ex/solve.hs
@@ -1,4 +1,4 @@
-solve :: Float -> Float
+solve :: Double -> Double
 solve x = sum $ 1 : [(x**y)/(product [1..y]) | y <- [1..9]]
 
 main :: IO()
